pyproject.toml
[project]
name = "csp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.7",
    "google-auth>=2.48.0",
    "google-cloud-firestore>=2.23.0",
    "google-cloud-storage>=3.9.0",
    "jinja2>=3.1.6",
    "pymongo[srv]>=4.16.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.2.1",
]
